requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
azure-cosmos==4.5.1
azure-identity==1.15.0
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.3
